+++ src/app/src/test/kotlin/org/imperial/mrc/hint/integration/adr/ADRTests.kt
@@ -1,10 +1,11 @@
 package org.imperial.mrc.hint.integration.adr
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import org.assertj.core.api.Assertions
 import org.assertj.core.api.Assertions.assertThat
+import org.imperial.mrc.hint.ConfiguredAppProperties
 import org.imperial.mrc.hint.helpers.JSONValidator
 import org.imperial.mrc.hint.integration.SecureIntegrationTests
+import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.EnumSource
@@ -59,6 +60,23 @@ class ADRTests : SecureIntegrationTests()
         {
             val data = ObjectMapper().readTree(result.body!!)["data"]
             assertThat(data["id"].textValue()).isEqualTo(id)
+        }
+    }
+
+    @ParameterizedTest
+    @EnumSource(IsAuthorized::class)
+    fun `can get orgs with permission`(isAuthorized: IsAuthorized)
+    {
+        testRestTemplate.postForEntity<String>("/adr/key", getPostEntityWithKey())
+        val result = testRestTemplate.getForEntity<String>("/adr/orgs?permission=update_dataset")
+
+        assertSecureWithSuccess(isAuthorized, result, null)
+
+        if (isAuthorized == IsAuthorized.TRUE)
+        {
+            val data = ObjectMapper().readTree(result.body!!)["data"]
+            assertThat(data.count()).isEqualTo(1)
+            assertThat(data[0]["name"].textValue()).isEqualTo("naomi-development-team")
         }
     }
 
@@ -164,6 +182,22 @@ class ADRTests : SecureIntegrationTests()
         assertSuccess(result, null)
         data = ObjectMapper().readTree(result.body!!)["data"]
         assertThat(data["pjnz"].textValue()).isEqualTo("inputs-unaids-spectrum-file")
+    }
+
+    @ParameterizedTest
+    @EnumSource(IsAuthorized::class)
+    fun `can push file to ADR`(isAuthorized: IsAuthorized)
+    {
+        if (isAuthorized == IsAuthorized.TRUE) {
+            val modelCalibrationId = waitForModelRunResult()
+            testRestTemplate.postForEntity<String>("/adr/key", getPostEntityWithKey())
+            val url = "/adr/datasets/hint_test/resource/${ConfiguredAppProperties().adrOutputZipSchema}/$modelCalibrationId?resourceFileName=output.zip"
+            val createResult = testRestTemplate.postForEntity<String>(url)
+            assertSuccess(createResult)
+            val resourceId = ObjectMapper().readTree(createResult.body!!)["data"]["id"].textValue()
+            val updateResult = testRestTemplate.postForEntity<String>("$url&resourceId=$resourceId")
+            assertSuccess(updateResult)
+        }
     }
 
     private fun getPostEntityWithKey(): HttpEntity<LinkedMultiValueMap<String, String>>