+++ src/app/static/src/tests/baseline/actions.test.ts
@@ -26,7 +26,9 @@ const adrSchemas: ADRSchemas = {
     shape: "shape",
     survey: "survey",
     programme: "program",
-    anc: "anc"
+    anc: "anc",
+    outputZip: "zip",
+    outputSummary: "summary"
 };
 
 const rootState = mockRootState({
@@ -432,13 +434,13 @@ describe("Baseline actions", () => {
         mockAxios.onGet("/adr/datasets/1234")
             .reply(200, mockSuccess({
                 resources: [
-                    {url: "something.com", last_modified: "2020-11-01", metadata_modified: "2020-11-02", resource_type: "pop"},
-                    {url: "something.com", last_modified: "2020-11-03", metadata_modified: "2020-11-04", resource_type: "pjnz"},
-                    {url: "something.com", last_modified: "2020-11-05", metadata_modified: "2020-11-06", resource_type: "shape"},
-                    {url: "something.com", last_modified: "2020-11-07", metadata_modified: "2020-11-08", resource_type: "survey"},
-                    {url: "something.com", last_modified: "2020-11-09", metadata_modified: "2020-11-10", resource_type: "program"},
-                    {url: "something.com", last_modified: "2020-11-11", metadata_modified: "2020-11-12", resource_type: "anc"},
-                    {url: "something.com", last_modified: "2020-10-01", metadata_modified: "2020-10-02", resource_type: "random"},
+                    {id: "1", url: "something.com", last_modified: "2020-11-01", metadata_modified: "2020-11-02", resource_type: "pop"},
+                    {id: "2", url: "something.com", last_modified: "2020-11-03", metadata_modified: "2020-11-04", resource_type: "pjnz"},
+                    {id: "3", url: "something.com", last_modified: "2020-11-05", metadata_modified: "2020-11-06", resource_type: "shape"},
+                    {id: "4", url: "something.com", last_modified: "2020-11-07", metadata_modified: "2020-11-08", resource_type: "survey"},
+                    {id: "5", url: "something.com", last_modified: "2020-11-09", metadata_modified: "2020-11-10", resource_type: "program"},
+                    {id: "6", url: "something.com", last_modified: "2020-11-11", metadata_modified: "2020-11-12", resource_type: "anc"},
+                    {id: "7", url: "something.com", last_modified: "2020-10-01", metadata_modified: "2020-10-02", resource_type: "random"},
 
                 ]
             }))
@@ -452,12 +454,12 @@ describe("Baseline actions", () => {
 
         expect(commit.mock.calls[0][0]).toBe(BaselineMutation.UpdateDatasetResources);
         expect(commit.mock.calls[0][1]).toEqual({
-            pjnz: mockDatasetResource({url: "something.com", lastModified: "2020-11-03", metadataModified: "2020-11-04",}),
-            shape: mockDatasetResource({url: "something.com", lastModified: "2020-11-05", metadataModified: "2020-11-06"}),
-            pop: mockDatasetResource({url: "something.com", lastModified: "2020-11-01", metadataModified: "2020-11-02"}),
-            survey: mockDatasetResource({url: "something.com", lastModified: "2020-11-07", metadataModified: "2020-11-08"}),
-            program: mockDatasetResource({url: "something.com", lastModified: "2020-11-09", metadataModified: "2020-11-10"}),
-            anc: mockDatasetResource({url: "something.com", lastModified: "2020-11-11", metadataModified: "2020-11-12",})
+            pjnz: mockDatasetResource({id: "2", url: "something.com", lastModified: "2020-11-03", metadataModified: "2020-11-04",}),
+            shape: mockDatasetResource({id: "3", url: "something.com", lastModified: "2020-11-05", metadataModified: "2020-11-06"}),
+            pop: mockDatasetResource({id: "1", url: "something.com", lastModified: "2020-11-01", metadataModified: "2020-11-02"}),
+            survey: mockDatasetResource({id: "4", url: "something.com", lastModified: "2020-11-07", metadataModified: "2020-11-08"}),
+            program: mockDatasetResource({id: "5", url: "something.com", lastModified: "2020-11-09", metadataModified: "2020-11-10"}),
+            anc: mockDatasetResource({id: "6", url: "something.com", lastModified: "2020-11-11", metadataModified: "2020-11-12",})
         });
     });