+++ src/app/static/src/tests/mocks.ts
@@ -368,12 +368,16 @@ export const mockDataset = (props: Partial<Dataset> = {}): Dataset => {
             shape: null,
             anc: null
         },
+        organization: {
+          id: "456"
+        },
         ...props
     }
 };
 
 export const mockDatasetResource = (props: Partial<DatasetResource> = {}): DatasetResource => {
     return {
+        id: "123",
         url: "www.something.com",
         lastModified: "2020-11-05T00:00:00",
         metadataModified: "2020-11-04T00:00:00",