+++ NEWS.md
@@ -1,6 +1,19 @@
-# hint 1.20.0
+
+# hint 1.23.0
 
 * add upload button and dialog box for result outputs to ADR
+
+# hint 1.22.0
+
+* Check if user has ADR upload permission
+
+# hint 1.21.0
+
+* Backend support for ADR uploads
+
+# hint 1.20.0
+
+* Add output adr schemas
 
 # hint 1.19.0